requirements.txt
# Computer Vision Leukocyte Classification Project
# Python 3.11+

# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# fastai library (required for project)
fastai>=2.7.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Image processing
Pillow>=10.0.0

# Machine learning utilities
scikit-learn>=1.3.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
